pyproject.toml
[project]
name = "lark-brief"
version = "0.1.0"
description = "每日资讯简报自动生成工具：爬取 → LLM 总结 → 格式化简报"
requires-python = ">=3.10"
dependencies = [
  "crawl4ai>=0.8.0",
  "openai>=1.0.0",
  "borax>=4.0.0",
  "pyyaml>=6.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "mypy>=1.0",
  "ruff>=0.8",
  "types-pyyaml>=6.0.12.20250915",
  "types-aiofiles>=25.1.0.20251011",
  "pytest-cov>=6.0",
]

[project.scripts]
lark-brief = "src.main:main"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501", "B008", "N802", "N818"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
